pyproject.toml
[project]
name = "Lubezki-Backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "boto3>=1.40.6",
    "boto3-stubs[full]>=1.40.6",
    "fastapi[standard]>=0.116.1",
    "google-genai>=1.28.0",
    "greenlet>=3.2.3",
    "mypy>=1.17.1",
    "pillow>=11.3.0",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.42",
    "sqlmodel>=0.0.24",
]

[tool.mypy]
plugins = ["pydantic.mypy"]
